requirements.txt
google-generativeai==0.1.0
gradio==3.5.0
requests==2.28.2
numpy==1.25.0
pandas==2.0.3
pytest==7.4.0
flake8==6.0.0
radon==6.0.1
python-dotenv
pygments
httpx==0.23.0
